refactor: UserEventMessage.java 위치 이동 -> 다른 도메인처럼 User 패키지 생성 후 이동
